Add penalty system for fitness calculation, Final fix for fitness values not updating on mutation
Add penalty system to fitness calculation for genes. A penalty is applied for each group in a gene where > 1/2 of the students are from the same faculty. Algorithm now takes in Faculty instead of Major for the /questiondata endpoint.

One final fix (hopefully) for issue where fitness values do not update correctly during swap mutations.